refactor(fsmeta): keep workloads in benchmark module
