+++ benchmark/cmd/fsmeta-demo/main.go
@@ -8,9 +8,9 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/feichai0017/NoKV/benchmark/fsmeta/workload"
 	"github.com/feichai0017/NoKV/fsmeta"
 	fsmetaclient "github.com/feichai0017/NoKV/fsmeta/client"
-	"github.com/feichai0017/NoKV/fsmeta/workload"
 )
 
 func main() {
